requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0

# Quantization (optional, for LLM inference)
bitsandbytes>=0.41.0

# Evaluation & Metrics
scikit-learn>=1.3.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Logging
tensorboard>=2.14.0

sentencepiece
protobuf